Expose client (browser) logs for integration tests running in browser (#180102) (#181513)
